Pass unsigned value into base::make_span in net
base::span will shortly require that length is unsigned and
fit into a size_t type.

This CL adds a checked_cast from a uint64_t to a size_t in
websocket channel, as on 32-bit this 64-bit value might exceed
the size_t value used by a base::span.

It also adds a long to size_t checked_cast in the root
store tool since `PEM_read_bio` returns a long which might
be unsigned.

This is part of a larger CL https://crrev.com/c/4211460.

BUG=1385166

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4214713
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1100091}

+++ net/tools/root_store_tool/root_store_tool.cc
@@ -14,6 +14,7 @@
 #include "base/files/file_path.h"
 #include "base/files/file_util.h"
 #include "base/logging.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/path_service.h"
 #include "base/strings/string_number_conversions.h"
 #include "base/strings/string_piece.h"
@@ -70,8 +71,9 @@ absl::optional<std::map<std::string, std::string>> DecodeCerts(
                  << " instead of CERTIFICATE";
       return absl::nullopt;
     }
-    std::string sha256_hex = base::ToLowerASCII(
-        base::HexEncode(crypto::SHA256Hash(base::make_span(data, len))));
+    std::string sha256_hex =
+        base::ToLowerASCII(base::HexEncode(crypto::SHA256Hash(
+            base::make_span(data, base::checked_cast<size_t>(len)))));
     certs[sha256_hex] = std::string(data, data + len);
   }
   return std::move(certs);
